# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodTinder API",
        "description": "API do logowania, swipe’ów i rekomendacji dań. Po prostu Tinder dla jedzenia!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/dishes": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Lista potraw",
                "operationId": "ddf841d9ba35d917bd22a504c13b2f72",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dish"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Utwórz nową potrawę",
                "operationId": "4055da700195ec26e05d953ac8fe8a0a",
                "requestBody": {
                    "description": "Użyj multipart/form-data, obrazek jest opcjonalny",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "cuisine_id",
                                    "flavour_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pizza Margherita"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cuisine_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "flavour_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Klasyk",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Obrazek (opcjonalnie). Zapisujesz w root dysku 'public'.",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utworzono",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dish"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dishes/{id}": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Pokaż jedną potrawę",
                "operationId": "07b6fe7c8dc79b35e66b58956b97b893",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dish"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Zaktualizuj potrawę",
                "operationId": "52b7975eeaea935e1251d215f5ac6bb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Wspiera JSON lub multipart/form-data. W JSON obrazka nie wyślesz.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "cuisine_id": {
                                        "type": "integer"
                                    },
                                    "flavour_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "remove_image": {
                                        "description": "Ustaw true, by usunąć istniejący obrazek",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "cuisine_id": {
                                        "type": "integer"
                                    },
                                    "flavour_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "remove_image": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dish"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Usuń potrawę",
                "operationId": "57bb2309cba8e67241d5d84d8ceb6beb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Dish": {
                "required": [
                    "id",
                    "name",
                    "category_id",
                    "cuisine_id",
                    "flavour_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "name": {
                        "type": "string",
                        "example": "Spaghetti Carbonara"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cuisine_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "flavour_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "image_url": {
                        "type": "string",
                        "example": "spaghetti.jpg",
                        "nullable": true
                    },
                    "image_url_full": {
                        "type": "string",
                        "example": "http://localhost/storage/spaghetti.jpg"
                    },
                    "description": {
                        "type": "string",
                        "example": "Classic Roman pasta.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}